@@ -21,8 +21,8 @@
     "@uniswap/v2-sdk": "^3.0.1",
     "blo": "^1.0.1",
     "daisyui": "^4.4.19",
-    "next": "^13.1.6",
-    "nextjs-progressbar": "^0.0.16",
+    "next": "^14.0.4",
+    "nprogress": "^0.2.0",
     "qrcode.react": "^3.1.0",
     "react": "^18.2.0",
     "react-copy-to-clipboard": "^5.1.0",
@@ -37,12 +37,13 @@
   "devDependencies": {
     "@trivago/prettier-plugin-sort-imports": "^4.1.1",
     "@types/node": "^17.0.35",
+    "@types/nprogress": "^0",
     "@types/react": "^18.0.9",
     "@types/react-copy-to-clipboard": "^5.0.4",
     "@typescript-eslint/eslint-plugin": "^5.39.0",
     "autoprefixer": "^10.4.12",
     "eslint": "^8.15.0",
-    "eslint-config-next": "^13.1.6",
+    "eslint-config-next": "^14.0.4",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.2.1",
     "postcss": "^8.4.16",
@@ -50,6 +51,6 @@
     "tailwindcss": "^3.3.3",
     "type-fest": "^4.6.0",
     "typescript": "^5.1.6",
-    "vercel": "^28.15.1"
+    "vercel": "^32.4.1"
   }
 }
