@@ -1,7 +1,9 @@
+"use client";
+
 import React, { useCallback, useRef, useState } from "react";
 import Image from "next/image";
 import Link from "next/link";
-import { useRouter } from "next/router";
+import { usePathname } from "next/navigation";
 import { Bars3Icon, BugAntIcon } from "@heroicons/react/24/outline";
 import { FaucetButton, RainbowKitCustomConnectButton } from "~~/components/scaffold-eth";
 import { useOutsideClick } from "~~/hooks/scaffold-eth";
@@ -25,12 +27,12 @@ export const menuLinks: HeaderMenuLink[] = [
 ];
 
 export const HeaderMenuLinks = () => {
-  const router = useRouter();
+  const pathname = usePathname();
 
   return (
     <>
       {menuLinks.map(({ label, href, icon }) => {
-        const isActive = router.pathname === href;
+        const isActive = pathname === href;
         return (
           <li key={href}>
             <Link
