@@ -1,6 +1,9 @@
+"use client";
+
 import { useEffect } from "react";
-import { useDarkMode, useIsMounted } from "usehooks-ts";
+import { useIsMounted } from "usehooks-ts";
 import { MoonIcon, SunIcon } from "@heroicons/react/24/outline";
+import { useDarkMode } from "~~/hooks/scaffold-eth/useDarkMode";
 
 export const SwitchTheme = ({ className }: { className?: string }) => {
   const { isDarkMode, toggle } = useDarkMode();
